requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Data processing
pillow>=9.0.0
opencv-python>=4.7.0
albumentations>=1.3.0

# Logging and visualization
tensorboard>=2.12.0
wandb>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.2.0
scipy>=1.10.0
